feat(home): create files component
